Finish test models move
