+++ tests/apps.py
@@ -1,0 +1,9 @@
+"""
+tests Django application initialization.
+"""
+
+from django.apps import AppConfig
+
+
+class TestsConfig(AppConfig):
+    name = "tests"
