+++ tests/forms.py
@@ -5,7 +5,7 @@ Forms to be used for testing.
 from __future__ import absolute_import, unicode_literals
 
 from edx_rbac.admin.forms import UserRoleAssignmentAdminForm
-from tests.test_models_app.models import ConcreteUserRoleAssignment
+from tests.models import ConcreteUserRoleAssignment
 
 
 class ConcreteUserRoleAssignmentAdminForm(UserRoleAssignmentAdminForm):
