+++ tests/models.py
@@ -6,8 +6,6 @@ These models exist solely for testing.
 They are not something that gets created when you install this application.
 """
 
-from __future__ import absolute_import, unicode_literals
-
 from edx_rbac.models import UserRole, UserRoleAssignment
