--- tests/test_models_app/__init__.py
@@ -1,3 +1,0 @@
-"""
-Django App to help tests the edx_rbac App.
-"""
