--- tests/test_models_app/apps.py
@@ -1,9 +1,0 @@
-"""
-test_models_app Django application initialization.
-"""
-
-from django.apps import AppConfig
-
-
-class TestModelsAppConfig(AppConfig):
-    name = 'test_models_app'
