+++ tox.ini
@@ -71,11 +71,11 @@ whitelist_externals =
 deps =
     -r{toxinidir}/requirements/quality.txt
 commands =
-    pylint edx_rbac tests/test_models_app test_utils manage.py setup.py
-    pylint --py3k edx_rbac tests/test_models_app test_utils manage.py setup.py
-    pycodestyle edx_rbac tests/test_models_app manage.py setup.py
-    pydocstyle edx_rbac tests/test_models_app manage.py setup.py
-    isort --check-only --diff --recursive tests/test_models_app test_utils edx_rbac manage.py setup.py test_settings.py
+    pylint edx_rbac tests test_utils manage.py setup.py
+    pylint --py3k edx_rbac tests test_utils manage.py setup.py
+    pycodestyle edx_rbac tests manage.py setup.py
+    pydocstyle edx_rbac tests manage.py setup.py
+    isort --check-only --diff --recursive tests test_utils edx_rbac manage.py setup.py test_settings.py
     make selfcheck
 
 [testenv:pii_check]
